+++ src/simpandas/__init__.py
@@ -5,16 +5,17 @@ Created on Sun Oct 11 11:14:32 2020
 @author: Martín Carlos Araya <martinaraya@gmail.com>
 """
 
-__version__ = '0.90.5'
-__release__ = 20260421
+__version__ = '0.90.6'
+__release__ = 20260425
 __all__ = ['SimSeries', 'SimDataFrame', 'ColumnUnits', 'read_excel', 'read_csv', 'read_json',
            'read_hdf5', 'read_summary', 'read_vdb', 'read_parquet',
-           'read_prodml', 'read_witsml', 'read_resqml', 'concat']
+           'read_prodml', 'read_witsml', 'read_resqml', 'read_auto', 'concat']
 
 from .series import SimSeries
 from .frame import SimDataFrame
 from .index import SimIndex
 from .readers import (read_excel, read_csv, read_json, read_hdf5, read_summary,
-                      read_vdb, read_parquet, read_prodml, read_witsml, read_resqml)
+                      read_vdb, read_parquet, read_prodml, read_witsml, read_resqml,
+                      read_auto)
 from .common.merger import concat
 from .common.units import ColumnUnits